Update patch set 1
Patch Set 1: Code-Review+1

> > This is mostly a dev setup unfamiliarity on my part, but why is
 > the
 > > dev set up not "automatic" during development?
 > >
 > > Running make build just seems like a bad route to go for this
 > type
 > > of project. I think there is just something wrong with the way
 > > static files are handled and if you also run Nginx during
 > > development instead of utilizing manage.py
 > >
 > > I think it would be better to fix whatever is causing the issue.
 > >
 > > See http://git.dpdk.org/tools/dpdk-ci-site/tree/docker-compose.yml#n41
 > > In this case "build" only needs to happen when new dependencies
 > are
 > > added. "migrate" when there are new migrations. But "dev" always
 > > just runs (ignoring livereload). So when changing templates,
 > python
 > > files, and static files, the developer doesn't need to do
 > anything.
 > 
 > I agree that none of this is ideal. I have struggled for a while to
 > come up with something better. Our static files live in a docker
 > volume so that they can be served by nginx. I can investigate a
 > solution without nginx using livereload, but I am not sure how well
 > your solution will translate because our docker environment is more
 > complicated (5+ containers).

Right. It may take a bit of changing things around. In a dev environment "collectstatic" shouldn't need to be ran, because manage.py would be serving the files (as long as you are using "static" tags correctly, as mentioned in a separate PR).
Ignore livereload for now, that is just a convenience middleware for also refreshing the web browser during development. 

I'll +1 this for now, but we can look into trying to utilize manage.py in a container maybe later this week or next week.

Patch-set: 1
Reviewer: Gerrit User 7509 <7509@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1