Update patch set 3
Patch Set 3:

(5 comments)

> Overall as we move state more into resources I'd like to see less
 > direct management of it in the API subsystem itself, just to keep
 > the code and the data it touches logically closer

I think its important to differentiate between the state of the resource and the state of the api jobs. I don't want to go too far in the other direction and have the resources responsible to managing the api and its state. I agree that we need to have a clean boundary between the two.

Patch-set: 3