Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)

66    for res in res_set:
67        iface_profile_set = InterfaceProfile.objects.filter(host=res.profile)
68        iface_set = Interface.objects.filter(Q(profile__in=iface_profile_set))
71        curr_profiles = [iface.profile for iface in iface_set]
72        missing_profiles = set(curr_profiles) - set(iface_profile_set)


Line 67 gets all the InterfaceProfile objects that are related to the current resource's profile. This line is correct.

Line 68 gets *all* interfaces who have one of the InterfaceProfiles from line 67. This will be dozens of interfaces, most of them not on the current resource. Since we are iterating over each resource, we should only check that resource's interfaces. I think this line should look like this:

68        iface_set = res.interfaces.all()


Line 71 is correct, but it can now be combined with line 68:

curr_profiles = [iface.profile for iface in res.interfaces.all()]


This line gives us all the InterfaceProfiles of all the Interfaces on the current resource. I think that is what you are after. Now, we can do the set comparison. Remember, the goal is to make sure that each resource has an Interface for each InterfaceProfile on its ResourceProfile, and no extra. In other words, curr_profiles and iface_profile_set should be equivalent. The classic way to see if two sets are the same is to check if they are subsets of each other:

a = set(curr_profiles)
b = set(iface_profile_set)
is_equal = a <= b and b <= a

The way you had the query on line 68 meant that it would always be equal to iface_profile_set, even if the database was wrong. In english, the query is something like "Give me all Interfaces who's profile is in iface_profile_set. Now, list the profiles of those interfaces". The result is always going to be the same, because that's how you constructed the iface_set in the first place.

Patch-set: 1
Reviewer: Gerrit User 6929 <6929@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>